server/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
motor==3.6.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
google-generativeai==0.8.3
faster-whisper==1.1.1
